ad_banlane_target_value表计算及查询业务,目标匹配表查询，责任关键指标查询，成本管控表查询，增创价值表查询

+++ web-admin/src/main/java/edu/jmu/oscm/mapper/ItemMapper.java
@@ -4,6 +4,7 @@ import edu.jmu.oscm.model.Item;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
+import java.math.BigInteger;
 import java.util.List;
 
 /**
@@ -32,7 +33,7 @@ public interface ItemMapper {
      * @param itemId
      * @return
      */
-    Item getItemByItemId(@Param("itemId") int itemId);
+    Item getItemByItemId(@Param("itemId") BigInteger itemId);
 
     /**
      * 清空item表
